+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-05-11 13:07:34
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-05-12 12:14:38
+// Commit: docs(theme): update readme and API contracts for theme
 
 class themeComponentState {
-  final String commitRef = '1049';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-05-11T13:07:34+05:00');
+  final String commitRef = '1062';
+  final String lastMessage = 'docs(theme): update readme and API contracts for theme';
+  final DateTime timestamp = DateTime.parse('2026-05-12T12:14:38+05:00');
 }